requirements.txt
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
bcrypt==4.0.1
blinker==1.6.3
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
construct==2.10.69
dj-database-url==2.1.0
Django==3.2.19
django-cors==0.1
django-cors-headers==4.3.0
django-restframework==0.0.1
djangorestframework==3.14.0
Flask==3.0.0
Flask-Login==0.6.2
flask-mysql-connector==1.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
numpy==1.26.1
packaging==23.1
pandas==2.1.2
paystack==1.5.0
Pillow==9.5.0
ply==3.11
protobuf==4.21.12
psycopg2==2.9.9
pyasn1==0.5.0
pycryptodomex==3.19.0
Pygments==2.16.1
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==68.2.2
simplejson==3.19.1
six==1.16.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
tensor==0.3.6
Twisted==23.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uuid==1.30
Werkzeug==3.0.0
whitenoise==6.5.0
zope.interface==6.1
